+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v0.0.91 (2026-03-30)
+
+* [Skip `init=False` fields in nuked dump path to match marshmallow behavior](https://github.com/anna-money/marshmallow-recipe/pull/294)
+
+
 ## v0.0.90 (2026-03-29)
 
 * [Fix datetime microseconds formatting for nuked](https://github.com/anna-money/marshmallow-recipe/pull/283)